Put function into separate header file
This is the start of a journey to learn how to manage multiple files. In the moment, I didn't set up the make file to compile everything, so this branch is not functional yet.
